Move components related to Match info into single folder
